requirements.txt
aiohttp==3.8.1
pydrive==1.3.1
transformers==4.17.0
torch<=1.11.0
pyyaml==5.3.1